Create change light intensit with comibined colour in cyclic order
This task is to write a code that can make the light on the microbit light up for a second and change color with a sequence of red, green, blue, magenta, and white. By using the three color the LED light can produce, red, green, and blue, we can light up magenta by light up blue and red.