Create consiiton like light intensity,  smiley face at LED
write a code that can do different reactions depending on the light intensity microbit sensed. Such as, show a smiling face and say, ”It’s sunny,” when the light intensity is bigger than 200, show a heart on the 5x5 LED display when the light intensity is between 100 and 200, and say, “It’s dark,” when the light intensity is smaller than 100.